# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Reservas medicas",
        "description": "API para gestionar las reservas medicas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Prefijo comun de todas las rutas del router"
        }
    ],
    "paths": {
...
        "/reservas": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener todas las reservas. Opcionalmente se pueden pasar filtros como query params (idprofesional, idpaciente, estado, fecha)",
                "operationId": "52c2ce0c552adf641f8704d8b0bbab23",
                "parameters": [
                    {
                        "name": "idprofesional",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "idpaciente",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Confirmada | Completada | No asistio | Cancelada",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha_reserva",
                        "in": "query",
                        "description": "YYYY, YYYY-MM o YYYY-MM-DD",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de reservas (opcionalmente filtrada)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RespuestaReserva"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "limit": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Filtro no permitido o valor invalido",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "Filtro no permitido"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Realizar una reserva",
                "operationId": "3966e75c6ab3ba2e8dfa6e51f49956cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CrearReservaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva realizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespuestaReserva"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay resultado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "No se encontro un profesional con ese id"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Reserva superpuesta",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "Este usuario ya tiene una reserva para esa misma fecha"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservas/mis-reservas": {
            "get": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Obtener las reservas de un usuario ya sea profesional o paciente",
                "operationId": "78daf52d0763d4c9ced94187227b0982",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de las reservas del usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RespuestaReserva"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "limit": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay reservas",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "No tienes reservas realizadas"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservas/{id}": {
            "patch": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Actualizar una reserva",
                "operationId": "e1cb3a31364a344f7cb63ce653daa5a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActualizarReservaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespuestaReserva"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos erroneos",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "La fecha no puede ser en el pasado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay resultado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "No se encontro una reserva con ese id"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservas/cancelar/{id}": {
            "put": {
                "tags": [
                    "Reservas"
                ],
                "summary": "Cancelar una reserva",
                "operationId": "7ace14fb68b931efc8d22f3d0da8f22c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva cancelada!",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "EXITO": "Reserva cancelada!"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cancelacion tardia",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "Solo se puede cancelar una reserva con Xhs de anticipacion"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "La reserva ya fue cancelada o completada",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "La reserva ya fue cancelada!"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "Error Interno del Servidor!"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ActualizarReservaRequest": {
                "properties": {
                    "fecha_reserva": {
                        "type": "string",
                        "example": "2026-03-15 18:45:00",
                        "nullable": true
                    },
                    "estado_reserva": {
                        "type": "string",
                        "example": "Cancelada",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CrearReservaRequest": {
                "required": [
                    "idProfesional",
                    "idPaciente",
                    "fecha_reserva"
                ],
                "properties": {
                    "idProfesional": {
                        "type": "integer",
                        "example": 100
                    },
                    "idPaciente": {
                        "type": "integer",
                        "example": 123
                    },
                    "fecha_reserva": {
                        "type": "string",
                        "example": "2026-03-15 18:45:00"
                    },
                    "estado": {
                        "type": "string",
                        "example": "Confirmada"
                    }
                },
                "type": "object"
            },
            "PacienteReserva": {
                "description": "Datos del paciente embebidos en una reserva. Lo minimo para renderizar una\nlista sin pedirle al cliente una consulta extra por cada fila.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 24
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Ana"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Torres"
                    }
                },
                "type": "object"
            },
            "ProfesionalReserva": {
                "description": "Datos del profesional embebidos en una reserva. Incluye la profesion porque\nes lo que la UI muestra junto al nombre en el listado del paciente.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Lucero"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Gimenez"
                    },
                    "profesion": {
                        "type": "string",
                        "example": "Pediatra"
                    }
                },
                "type": "object"
            },
            "RespuestaReserva": {
                "description": "Forma unica de una reserva en la API: los participantes vienen anidados y ya\nresueltos por JOIN, para que el cliente no tenga que pedirlos aparte.",
                "properties": {
                    "idReserva": {
                        "type": "integer",
                        "example": 15
                    },
                    "paciente": {
                        "$ref": "#/components/schemas/PacienteReserva"
                    },
                    "profesional": {
                        "$ref": "#/components/schemas/ProfesionalReserva"
                    },
                    "fecha_reserva": {
                        "type": "string",
                        "example": "2026-04-18 15:00:00"
                    },
                    "estado": {
                        "type": "string",
                        "example": "Confirmada"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT aquí para autenticarte",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}